Add files via upload
This file contains all of the comments that were extracted from the top posts on R/Conspiracy and R/News in April 2021. The category is 1 if the post is found in R/News and is 2 if the post is found in R/Conspiracy. Note that his texts exist in raw form and have not yet been processed for text analysis.